GHCi.BreakArray

Description

Break Arrays

An array of words, indexed by a breakpoint number (breakpointId in Tickish) containing the ignore count for every breakpopint. There is one of these arrays per module.

For each word with value n: n > 1 : the corresponding breakpoint is enabled. Next time the bp is hit, GHCi will decrement the ignore count and continue processing. n == 0 : The breakpoint is enabled, GHCi will stop next time it hits this breakpoint. n == -1: This breakpoint is disabled. n < -1 : Not used.

Documentation

data BreakArray

Constructors

BA (MutableByteArray# RealWorld)

newBreakArray :: Int -> IO BreakArray

getBreak :: BreakArray -> Int -> IO (Maybe Int)

setupBreakpoint :: BreakArray -> Int -> Int -> IO Bool

breakOn :: Int

breakOff :: Int

showBreakArray :: BreakArray -> IO ()