GHCi.Debugger

Single step mode

rts_enableStopNextBreakpoint :: ThreadId# -> IO ()

Enables the single step mode for a specific thread, thus stopping only on breakpoints in that thread.

rts_enableStopNextBreakpointAll :: IO ()

Enables the single step mode for all threads, thus stopping at any existing breakpoint.

rts_disableStopNextBreakpoint :: ThreadId# -> IO ()

Disables per-thread single-step mode. Note: if global single-step is enabled we stop at all breakpoints regardless of the per-thread flag.

rts_disableStopNextBreakpointAll :: IO ()

Disables the single step mode for all threads

Step out mode

rts_enableStopAfterReturn :: ThreadId# -> IO ()

rts_disableStopAfterReturn :: ThreadId# -> IO ()

Stop on exception

exceptionFlag :: Ptr CInt

Breakpoint Callback

type BreakpointCallback = Addr# -> Addr# -> Int# -> Bool -> HValue -> IO ()

breakPointIOAction :: Ptr (StablePtr BreakpointCallback)