GHCi.InfoTable

Description

Run-time info table support. This module provides support for creating and reading info tables in the running program. We use the RTS data structures directly via hsc2hs.

Documentation

peekItbl :: Ptr StgInfoTable -> IO StgInfoTable

data StgInfoTable

Constructors

StgInfoTable

Fields

entry :: Maybe EntryFunPtr

ptrs :: HalfWord

nptrs :: HalfWord

tipe :: HalfWord

srtlen :: HalfWord

code :: Maybe ItblCodes

conInfoPtr :: Ptr () -> Ptr ()

Convert a pointer to an StgConInfo into an info pointer that can be used in the header of a closure.

mkConInfoTable :: Int -> Int -> Int -> Int -> [Word8] -> IO (Ptr StgInfoTable)