GHCi.Message

Description

Remote GHCi message types and serialization.

For details on Remote GHCi, see Note [Remote GHCi] in compilerGHCRuntime/Interpreter.hs.

Documentation

data Message a where

A Message a is a message that returns a value of type a. These are requests sent from GHC to the server.

Constructors

Shutdown :: Message ()
Exit the iserv process

RtsRevertCAFs :: Message ()

InitLinker :: Message ()

LookupSymbol :: String -> Message (Maybe (RemotePtr ()))

LookupClosure :: String -> Message (Maybe HValueRef)

LoadDLL :: String -> Message (Maybe String)

LoadArchive :: String -> Message ()

LoadObj :: String -> Message ()

UnloadObj :: String -> Message ()

AddLibrarySearchPath :: String -> Message (RemotePtr ())

RemoveLibrarySearchPath :: RemotePtr () -> Message Bool

ResolveObjs :: Message Bool

FindSystemLibrary :: String -> Message (Maybe String)

CreateBCOs :: [ByteString] -> Message [HValueRef]
Create a set of BCO objects, and return HValueRefs to them Note: Each ByteString contains a Binary-encoded [ResolvedBCO], not a ResolvedBCO. The list is to allow us to serialise the ResolvedBCOs in parallel. See createBCOs in compilerGHCRuntime/Interpreter.hs.

FreeHValueRefs :: [HValueRef] -> Message ()
Release HValueRefs

AddSptEntry :: Fingerprint -> HValueRef -> Message ()
Add entries to the Static Pointer Table

MallocData :: ByteString -> Message (RemotePtr ())
Malloc some data and return a RemotePtr to it

MallocStrings :: [ByteString] -> Message [RemotePtr ()]

PrepFFI :: FFIConv -> [FFIType] -> FFIType -> Message (RemotePtr C_ffi_cif)
Calls prepareForeignCall

FreeFFI :: RemotePtr C_ffi_cif -> Message ()
Free data previously created by PrepFFI

MkConInfoTable :: Bool -> Int -> Int -> Int -> Int -> ByteString -> Message (RemotePtr StgInfoTable)
Create an info table for a constructor

EvalStmt :: EvalOpts -> EvalExpr HValueRef -> Message (EvalStatus [HValueRef])
Evaluate a statement

ResumeStmt :: EvalOpts -> RemoteRef (ResumeContext [HValueRef]) -> Message (EvalStatus [HValueRef])
Resume evaluation of a statement after a breakpoint

AbandonStmt :: RemoteRef (ResumeContext [HValueRef]) -> Message ()
Abandon evaluation of a statement after a breakpoint

EvalString :: HValueRef -> Message (EvalResult String)
Evaluate something of type IO String

EvalStringToString :: HValueRef -> String -> Message (EvalResult String)
Evaluate something of type String -> IO String

EvalIO :: HValueRef -> Message (EvalResult ())
Evaluate something of type IO ()

MkCostCentres :: String -> [(String, String)] -> Message [RemotePtr CostCentre]
Create a set of CostCentres with the same module name

CostCentreStackInfo :: RemotePtr CostCentreStack -> Message [String]
Show a CostCentreStack as a [String]

NewBreakArray :: Int -> Message (RemoteRef BreakArray)
Create a new array of breakpoint flags

SetupBreakpoint :: RemoteRef BreakArray -> Int -> Int -> Message ()
Set how many times a breakpoint should be ignored also used for enable/disable

BreakpointStatus :: RemoteRef BreakArray -> Int -> Message Bool
Query the status of a breakpoint (True = enabled)

GetBreakpointVar :: HValueRef -> Int -> Message (Maybe HValueRef)
Get a reference to a free variable at a breakpoint

StartTH :: Message (RemoteRef (IORef QState))
Start a new TH module, return a state token that should be

RunTH :: RemoteRef (IORef QState) -> HValueRef -> THResultType -> Maybe Loc -> Message (QResult ByteString)
Evaluate a TH computation.

Returns a ByteString, because we have to force the result before returning it to ensure there are no errors lurking in it. The TH types don't have NFData instances, and even if they did, we have to serialize the value anyway, so we might as well serialize it to force it.

RunModFinalizers :: RemoteRef (IORef QState) -> [RemoteRef (Q ())] -> Message (QResult ())
Run the given mod finalizers.

GetClosure :: HValueRef -> Message (GenClosure HValueRef)
Remote interface to GHC.Exts.Heap.getClosureData. This is used by the GHCi debugger to inspect values in the heap for :print and type reconstruction.

Seq :: HValueRef -> Message (EvalStatus ())
Evaluate something. This is used to support :force in GHCi.

ResumeSeq :: RemoteRef (ResumeContext ()) -> Message (EvalStatus ())
Resume forcing a free variable in a breakpoint (#2950)

data Msg

Constructors

forall a.(Binary a, Show a) => Msg (Message a)

data THMessage a where

Messages sent back to GHC from GHCi.TH, to implement the methods of Quasi. For an overview of how TH works with Remote GHCi, see Note [Remote Template Haskell] in GHCi.TH.

Constructors

NewName :: String -> THMessage (THResult Name)

Report :: Bool -> String -> THMessage (THResult ())

LookupName :: Bool -> String -> THMessage (THResult (Maybe Name))

Reify :: Name -> THMessage (THResult Info)

ReifyFixity :: Name -> THMessage (THResult (Maybe Fixity))

ReifyType :: Name -> THMessage (THResult Type)

ReifyInstances :: Name -> [Type] -> THMessage (THResult [Dec])

ReifyRoles :: Name -> THMessage (THResult [Role])

ReifyAnnotations :: AnnLookup -> TypeRep -> THMessage (THResult [ByteString])

ReifyModule :: Module -> THMessage (THResult ModuleInfo)

ReifyConStrictness :: Name -> THMessage (THResult [DecidedStrictness])

AddDependentFile :: FilePath -> THMessage (THResult ())

AddTempFile :: String -> THMessage (THResult FilePath)

AddModFinalizer :: RemoteRef (Q ()) -> THMessage (THResult ())

AddCorePlugin :: String -> THMessage (THResult ())

AddTopDecls :: [Dec] -> THMessage (THResult ())

AddForeignFilePath :: ForeignSrcLang -> FilePath -> THMessage (THResult ())

IsExtEnabled :: Extension -> THMessage (THResult Bool)

ExtsEnabled :: THMessage (THResult [Extension])

PutDoc :: DocLoc -> String -> THMessage (THResult ())

GetDoc :: DocLoc -> THMessage (THResult (Maybe String))

StartRecover :: THMessage ()

EndRecover :: Bool -> THMessage ()

FailIfErrs :: THMessage (THResult ())

RunTHDone :: THMessage ()
Indicates that this RunTH is finished, and the next message will be the result of RunTH (a QResult).

data THMsg

Constructors

forall a.(Binary a, Show a) => THMsg (THMessage a)

data QResult a

Template Haskell return values

Constructors

QDone a
RunTH finished successfully; return value follows

QException String
RunTH threw an exception

QFail String
RunTH called fail

data EvalStatus_ a b

Constructors

EvalComplete Word64 (EvalResult a)

EvalBreak Bool HValueRef Int Int (RemoteRef (ResumeContext b)) (RemotePtr CostCentreStack)

type EvalStatus a = EvalStatus_ a a

data EvalResult a

Constructors

EvalException SerializableException

EvalSuccess a

data EvalOpts

Constructors

EvalOpts

Fields

useSandboxThread :: Bool

singleStep :: Bool

breakOnException :: Bool

breakOnError :: Bool

data EvalExpr a

We can pass simple expressions to EvalStmt, consisting of values and application. This allows us to wrap the statement to be executed in another function, which is used by GHCi to implement :set args and :set prog. It might be worthwhile to extend this little language in the future.

Constructors

EvalThis a

EvalApp (EvalExpr a) (EvalExpr a)

data SerializableException

Constructors

EUserInterrupt

EExitCode ExitCode

EOtherException String

toSerializableException :: SomeException -> SerializableException

fromSerializableException :: SerializableException -> SomeException

data THResult a

Constructors

THException String

THComplete a

data THResultType

Constructors

THExp

THPat

THType

THDec

THAnnWrapper

data ResumeContext a

Constructors

ResumeContext

Fields

resumeBreakMVar :: MVar ()

resumeStatusMVar :: MVar (EvalStatus a)

resumeThreadId :: ThreadId

data QState

The server-side Template Haskell state. This is created by the StartTH message. A new one is created per module that GHC typechecks.

Constructors

QState

Fields

qsMap :: Map TypeRep Dynamic

persistent data between splices in a module

qsLocation :: Maybe Loc

location for current splice, if any

qsPipe :: Pipe

pipe to communicate with GHC

getMessage :: Get Msg

putMessage :: Message a -> Put

getTHMessage :: Get THMsg

putTHMessage :: THMessage a -> Put

data Pipe

Constructors

Pipe

Fields

pipeRead :: Handle

pipeWrite :: Handle

pipeLeftovers :: IORef (Maybe ByteString)

remoteCall :: Binary a => Pipe -> Message a -> IO a

remoteTHCall :: Binary a => Pipe -> THMessage a -> IO a

readPipe :: Pipe -> Get a -> IO a

writePipe :: Pipe -> Put -> IO ()

Orphan instances

Binary ExitCode

Instance details

Methods

put :: ExitCode -> Put

get :: Get ExitCode

putList :: [ExitCode] -> Put

Binary ClosureType

Instance details

Methods

put :: ClosureType -> Put

get :: Get ClosureType

putList :: [ClosureType] -> Put

Binary PrimType

Instance details

Methods

put :: PrimType -> Put

get :: Get PrimType

putList :: [PrimType] -> Put

Binary TsoFlags

Instance details

Methods

put :: TsoFlags -> Put

get :: Get TsoFlags

putList :: [TsoFlags] -> Put

Binary WhatNext

Instance details

Methods

put :: WhatNext -> Put

get :: Get WhatNext

putList :: [WhatNext] -> Put

Binary WhyBlocked

Instance details

Methods

put :: WhyBlocked -> Put

get :: Get WhyBlocked

putList :: [WhyBlocked] -> Put

Binary StgInfoTable

Instance details

Methods

put :: StgInfoTable -> Put

get :: Get StgInfoTable

putList :: [StgInfoTable] -> Put

Binary CostCentre

Instance details

Methods

put :: CostCentre -> Put

get :: Get CostCentre

putList :: [CostCentre] -> Put

Binary CostCentreStack

Instance details

Methods

put :: CostCentreStack -> Put

get :: Get CostCentreStack

putList :: [CostCentreStack] -> Put

Binary IndexTable

Instance details

Methods

put :: IndexTable -> Put

get :: Get IndexTable

putList :: [IndexTable] -> Put

Binary StgTSOProfInfo

Instance details

Methods

put :: StgTSOProfInfo -> Put

get :: Get StgTSOProfInfo

putList :: [StgTSOProfInfo] -> Put

Binary (FunPtr a)

Instance details

Methods

put :: FunPtr a -> Put

get :: Get (FunPtr a)

putList :: [FunPtr a] -> Put

Binary (Ptr a)

Instance details

Methods

put :: Ptr a -> Put

get :: Get (Ptr a)

putList :: [Ptr a] -> Put

Binary a => Binary (GenClosure a)

Instance details

Methods

put :: GenClosure a -> Put

get :: Get (GenClosure a)

putList :: [GenClosure a] -> Put