GHCi.ObjLink

Description

Primarily, this module consists of an interface to the C-land dynamic linker.

Documentation

initObjLinker :: ShouldRetainCAFs -> IO ()

data ShouldRetainCAFs

Constructors

RetainCAFs
Retain CAFs unconditionally in linked Haskell code. Note that this prevents any code from being unloaded. It should not be necessary unless you are GHCi or hs-plugins, which needs to be able call any function in the compiled code.

DontRetainCAFs
Do not retain CAFs. Everything reachable from foreign exports will be retained, due to the StablePtrs created by the module initialisation code. unloadObj frees these StablePtrs, which will allow the CAFs to be GC'd and the code to be removed.

loadDLL :: String -> IO (Maybe String)

loadDLL loads a dynamic library using the OS's native linker (i.e. dlopen() on Unix, LoadLibrary() on Windows). It takes either an absolute pathname to the file, or a relative filename (e.g. "libfoo.so" or "foo.dll"). In the latter case, loadDLL searches the standard locations for the appropriate library.

loadArchive :: String -> IO ()

loadObj :: String -> IO ()

unloadObj :: String -> IO ()

unloadObj drops the given dynamic library from the symbol table as well as enables the library to be removed from memory during a future major GC.

purgeObj :: String -> IO ()

purgeObj drops the symbols for the dynamic library from the symbol table. Unlike unloadObj, the library will not be dropped memory during a future major GC.

lookupSymbol :: String -> IO (Maybe (Ptr a))

lookupClosure :: String -> IO (Maybe HValueRef)

resolveObjs :: IO Bool

addLibrarySearchPath :: String -> IO (Ptr ())

removeLibrarySearchPath :: Ptr () -> IO Bool

findSystemLibrary :: String -> IO (Maybe String)