GHCi.RemoteTypes

Description

Types for referring to remote objects in Remote GHCi. For more details, see Note [External GHCi pointers] in compilerGHCRuntime/Interpreter.hs

For details on Remote GHCi, see Note [Remote GHCi] in compilerGHCRuntime/Interpreter.hs.

Documentation

newtype RemotePtr a

Constructors

RemotePtr Word64

toRemotePtr :: Ptr a -> RemotePtr a

fromRemotePtr :: RemotePtr a -> Ptr a

castRemotePtr :: RemotePtr a -> RemotePtr b

newtype HValue

Constructors

HValue Any

data RemoteRef a

A reference to a remote value. These are allocated and freed explicitly.

mkRemoteRef :: a -> IO (RemoteRef a)

Make a reference to a local value that we can send remotely. This reference will keep the value that it refers to alive until freeRemoteRef is called.

localRef :: RemoteRef a -> IO a

Convert an HValueRef to an HValue. Should only be used if the HValue originated in this process.

freeRemoteRef :: RemoteRef a -> IO ()

Release an HValueRef that originated in this process

type HValueRef = RemoteRef HValue

toHValueRef :: RemoteRef a -> RemoteRef HValue

data ForeignRef a

An HValueRef with a finalizer

mkForeignRef :: RemoteRef a -> IO () -> IO (ForeignRef a)

Create a ForeignRef from a RemoteRef. The finalizer should arrange to call freeHValueRef on the HValueRef. (since this function needs to be called in the process that created the HValueRef, it cannot be called directly from the finalizer).

withForeignRef :: ForeignRef a -> (RemoteRef a -> IO b) -> IO b

Use a ForeignHValue

type ForeignHValue = ForeignRef HValue

unsafeForeignRefToRemoteRef :: ForeignRef a -> RemoteRef a

finalizeForeignRef :: ForeignRef a -> IO ()