GHCi.ResolvedBCO

Documentation

data ResolvedBCO

A ResolvedBCO is one in which all the Name references have been resolved to actual addresses or RemoteHValues.

Constructors

ResolvedBCO

Fields

resolvedBCOIsLE :: Bool

resolvedBCOArity :: !Int

resolvedBCOInstrs :: BCOByteArray Word16

resolvedBCOBitmap :: BCOByteArray Word

resolvedBCOLits :: BCOByteArray Word

resolvedBCOPtrs :: SizedSeq ResolvedBCOPtr

data ResolvedBCOPtr

Constructors

ResolvedBCORef !Int
reference to the Nth BCO in the current set

ResolvedBCOPtr !(RemoteRef HValue)
reference to a previously created BCO

ResolvedBCOStaticPtr !(RemotePtr ())
reference to a static ptr

ResolvedBCOPtrBCO ResolvedBCO
a nested BCO

ResolvedBCOPtrBreakArray !(RemoteRef BreakArray)
Resolves to the MutableArray# inside the BreakArray

isLittleEndian :: Bool

data BCOByteArray a

Wrapper for a ByteArray#. The phantom type tells what elements are stored in the ByteArray#. Creating a ByteArray# can be achieved using UArray's API, where the underlying ByteArray# can be unpacked.

Constructors

BCOByteArray

Fields

getBCOByteArray :: !ByteArray#

mkBCOByteArray :: UArray Int a -> BCOByteArray a