GHCi.ObjLink

Description

Primarily, this module consists of an interface to the C-land dynamic linker.

Documentation

initObjLinker :: IO ()

loadDLL :: String -> IO (Maybe String)

loadDLL loads a dynamic library using the OS's native linker (i.e. dlopen() on Unix, LoadLibrary() on Windows). It takes either an absolute pathname to the file, or a relative filename (e.g. "libfoo.so" or "foo.dll"). In the latter case, loadDLL searches the standard locations for the appropriate library.

loadArchive :: String -> IO ()

loadObj :: String -> IO ()

unloadObj :: String -> IO ()

lookupSymbol :: String -> IO (Maybe (Ptr a))

lookupClosure :: String -> IO (Maybe HValueRef)

resolveObjs :: IO Bool

addLibrarySearchPath :: String -> IO (Ptr ())

removeLibrarySearchPath :: Ptr () -> IO Bool

findSystemLibrary :: String -> IO (Maybe String)