GHCi.ResolvedBCO

Documentation

data ResolvedBCO

A ResolvedBCO is one in which all the Name references have been resolved to actual addresses or RemoteHValues.

Note, all arrays are zero-indexed (we assume this when serializing/deserializing)

Constructors

ResolvedBCO

Fields

resolvedBCOIsLE :: Bool

resolvedBCOArity :: !Int

resolvedBCOInstrs :: UArray Int Word16

resolvedBCOBitmap :: UArray Int Word64

resolvedBCOLits :: UArray Int Word64

resolvedBCOPtrs :: SizedSeq ResolvedBCOPtr

data ResolvedBCOPtr

Constructors

ResolvedBCORef !Int
reference to the Nth BCO in the current set

ResolvedBCOPtr !(RemoteRef HValue)
reference to a previously created BCO

ResolvedBCOStaticPtr !(RemotePtr ())
reference to a static ptr

ResolvedBCOPtrBCO ResolvedBCO
a nested BCO

ResolvedBCOPtrBreakArray !(RemoteRef BreakArray)
Resolves to the MutableArray# inside the BreakArray

isLittleEndian :: Bool