GHCi.Signals

Documentation

installSignalHandlers :: IO ()

Install standard signal handlers for catching ^C, which just throw an exception in the target thread. The current target thread is the thread at the head of the list in the MVar passed to installSignalHandlers.